improve for merge
