+++ tasks/syllogism-reasoning.md
@@ -29,13 +29,365 @@ Socrates is mortal.
 Input:
 
 ```
-All fruits have seeds. An apple is a fruit. What can we conclude?
+No social media platforms are completely private. Instagram is a social media platform. What can we conclude?
 ```
 
 Output:
 
 ```
-An apple has seeds.
+Instagram is not completely private.
+```
+
+---
+
+### Example 3:
+
+Input:
+
+```
+All smartphones are internet-capable devices. Some handheld gadgets are smartphones. What can we conclude?
+```
+
+Output:
+
+```
+Some handheld gadgets are internet-capable devices.
+```
+
+---
+
+### Example 4:
+
+Input:
+
+```
+No fast-food meals are nutritionally balanced. Some popular lunches are fast-food meals. What can we conclude?
+```
+
+Output:
+
+```
+Some popular lunches are not nutritionally balanced.
+```
+
+---
+
+### Example 5:
+
+Input:
+
+```
+No renewable energy sources produce greenhouse gases. Coal power plants produce greenhouse gases. What can we conclude?
+```
+
+Output:
+
+```
+Coal power plants are not renewable energy sources.
+```
+
+---
+
+### Example 6:
+
+Input:
+
+```
+All solar panels convert sunlight to electricity. No wind turbines convert sunlight to electricity. What can we conclude?
+```
+
+Output:
+
+```
+No wind turbines are solar panels.
+```
+
+---
+
+### Example 7:
+
+Input:
+
+```
+No organic foods contain artificial preservatives. Some popular snacks contain artificial preservatives. What can we conclude?
+```
+
+Output:
+
+```
+Some popular snacks are not organic foods.
+```
+
+---
+
+### Example 8:
+
+Input:
+
+```
+All college graduates have a degree. Some job applicants do not have a degree. What can we conclude?
+```
+
+Output:
+
+```
+Some job applicants are not college graduates.
+```
+
+---
+
+### Example 9:
+
+Input:
+
+```
+All hybrid cars are fuel-efficient. All hybrid cars have electric motors. What can we conclude?
+```
+
+Output:
+
+```
+Some vehicles with electric motors are fuel-efficient.
+```
+
+---
+
+### Example 10:
+
+Input:
+
+```
+Some online courses are free. All online courses are accessible from home. What can we conclude?
+```
+
+Output:
+
+```
+Some courses accessible from home are free.
+```
+
+---
+
+### Example 11:
+
+Input:
+
+```
+All smartphones have touchscreens. Some smartphones are budget-friendly. What can we conclude?
+```
+
+Output:
+
+```
+Some budget-friendly devices have touchscreens.
+```
+
+---
+
+### Example 12:
+
+Input:
+
+```
+No fossil fuels are sustainable energy sources. All fossil fuels are carbon-emitting. What can we conclude?
+```
+
+Output:
+
+```
+Some carbon-emitting energy sources are not sustainable.
+```
+
+---
+
+### Example 13:
+
+Input:
+
+```
+Some best-selling books are not fiction. All best-selling books are widely read. What can we conclude?
+```
+
+Output:
+
+```
+Some widely read books are not fiction.
+```
+
+---
+
+### Example 14:
+
+Input:
+
+```
+No single-use plastics are environmentally friendly. Some packaging materials are single-use plastics. What can we conclude?
+```
+
+Output:
+
+```
+Some packaging materials are not environmentally friendly.
+```
+
+---
+
+### Example 15:
+
+Input:
+
+```
+All electric vehicles have rechargeable batteries. No gasoline-powered cars have rechargeable batteries. What can we conclude?
+```
+
+Output:
+
+```
+No gasoline-powered cars are electric vehicles.
+```
+
+---
+
+### Example 16:
+
+Input:
+
+```
+Some renewable energy sources are wind power. All renewable energy sources are environmentally friendly. What can we conclude?
+```
+
+Output:
+
+```
+Some environmentally friendly energy sources are wind power.
+```
+
+---
+
+### Example 17:
+
+Input:
+
+```
+No junk foods are nutritious. All potato chips are junk foods. What can we conclude?
+```
+
+Output:
+
+```
+Some potato chips are not nutritious.
+```
+
+---
+
+### Example 18:
+
+Input:
+
+```
+No digital currencies are physical coins. All physical coins are tangible assets. What can we conclude?
+```
+
+Output:
+
+```
+Some tangible assets are not digital currencies.
+```
+
+---
+
+### Example 19:
+
+Input:
+
+```
+All vegetarian dishes are meat-free. No dishes containing beef are meat-free. What can we conclude?
+```
+
+Output:
+
+```
+No dishes containing beef are vegetarian dishes.
+```
+
+---
+
+### Example 20:
+
+Input:
+
+```
+No violent video games are educational. Some popular games are violent video games. What can we conclude?
+```
+
+Output:
+
+```
+Some popular games are not educational.
+```
+
+---
+
+### Example 21:
+
+Input:
+
+```
+All smartphones are portable devices. All portable devices are battery-powered. What can we conclude?
+```
+
+Output:
+
+```
+Some battery-powered items are smartphones.
+```
+
+---
+
+### Example 22:
+
+Input:
+
+```
+All electric cars are zero-emission vehicles. No diesel engines are zero-emission vehicles. What can we conclude?
+```
+
+Output:
+
+```
+No diesel engines are electric cars.
+```
+
+---
+
+### Example 23:
+
+Input:
+
+```
+All hummingbirds are nectar-feeding birds. All hummingbirds can hover in mid-air. What can we conclude?
+```
+
+Output:
+
+```
+Some birds that can hover in mid-air are nectar-feeding birds.
+```
+
+---
+
+### Example 24:
+
+Input:
+
+```
+No open-source software requires paid licenses. All open-source software is modifiable by users. What can we conclude?
+```
+
+Output:
+
+```
+Some software that is modifiable by users does not require paid licenses.
 ```
 
 ## Tags:
